Updated project
Corrected the issue that dealt with the relationship between users and students/instructors entities (FK on user_id). When registering a new student/instructor, the current user_id inserted as well.

Upon successfully registering as a student/instructor, user is assigned that role type. 

Added validation for accessing the student/instructor registration forms. (Error message pops up if user_id already exists in one of the tables)

On the table listing the users, the "edit" link is now only visible to a user with the role of Admin.